day 6

+++ 2023/day06/test.h
@@ -4,6 +4,6 @@
 
 namespace aoc2023_day06 {
     int part_1(std::string_view path);
-    int part_2(std::string_view path);
+    unsigned long long part_2(std::string_view path);
 }
